[OPIK-8198] [FE] alerts: rework Enable alert and webhook testing placement
- Hide the "Enable alert" toggle when creating an alert; new alerts still
  default to enabled. On edit it moves to the bottom of the form, directly
  above the submit buttons.
- Drop the "Test alert configuration" side panel. "Test connection" now sits
  on the Endpoint URL row in Webhook settings.
- Move the per-trigger payload example and "Test trigger" button into each
  trigger row, collapsed behind an "Example payload" toggle.
- Extract the shared test/validation logic into useWebhookTest so both entry
  points share one mutation and the same toast handling.

The form is now single-column; the sticky right-hand column it existed for
is gone.
